feat(event-manager): reflect loaded database state
