+++ src/renderer/src/features/EventManager/hooks/useEventDatabases.ts
@@ -2,11 +2,12 @@ import { useQuery } from "@tanstack/react-query";
 import { useIpcRenderer } from "~/hooks/useIpcRenderer";
 import { EventDatabaseMetadata } from "$shared/models";
 
-export function useEventDatabases() {
+export function useEventDatabases(enabled = true) {
   const ipcRenderer = useIpcRenderer();
 
   return useQuery({
     queryKey: ["event-databases"],
+    enabled,
     queryFn: async (): Promise<EventDatabaseMetadata[]> => {
       const result = (await ipcRenderer.invoke("list-event-databases")) as EventDatabaseMetadata[];
       return result || [];
@@ -27,11 +28,23 @@ export function useActiveDatabaseSlug() {
   });
 }
 
-export function useEventDatabaseBackups() {
+export function useIsEventDatabaseLoaded() {
+  const ipcRenderer = useIpcRenderer();
+
+  return useQuery({
+    queryKey: ["is-event-database-loaded"],
+    queryFn: async (): Promise<boolean> => {
+      return (await ipcRenderer.invoke("is-event-database-loaded")) as boolean;
+    }
+  });
+}
+
+export function useEventDatabaseBackups(enabled = true) {
   const ipcRenderer = useIpcRenderer();
 
   return useQuery({
     queryKey: ["event-database-backups"],
+    enabled,
     queryFn: async (): Promise<EventDatabaseMetadata[]> => {
       const result = (await ipcRenderer.invoke(
         "list-event-database-backups"
